Change RX switch behavior to protect L293D
There was a problem pointed out that the L293D pin 16
must receive at least as much power as it receives into
any of its input pins. One of the switches was wired in
a way that it could violate this rule, and potentially
damage the chip. Since the purpose of the rightmost switch
was to turn the radio receiver on or off, this switch and
the RX was rewired. Now, L293D will continuously receive
5 volts at pin 16 whenever the Arduino board is on.